refine vectorizer code and config
