+++ kag/examples/hotpotqa/kag_config.cfg
@@ -4,7 +4,7 @@ host_addr = http://127.0.0.1:8887
 id = 4
 
 [vectorizer]
-vectorizer = knext.common.vectorizer.OpenAIVectorizer
+vectorizer = kag.common.vectorizer.OpenAIVectorizer
 model = bge-m3
 api_key = EMPTY
 base_url = http://127.0.0.1:11434/v1
