+++ tests/solver/logic_form_executor/kag_config.cfg
@@ -2,7 +2,7 @@
 id = 4
 
 [vectorizer]
-vectorizer = knext.common.vectorizer.OpenAIVectorizer
+vectorizer = kag.common.vectorizer.OpenAIVectorizer
 model = bge-m3
 api_key = EMPTY
 base_url = http://127.0.0.1:11434/v1
