+++ tests/solver/logic_form_plan/kag_config.cfg
@@ -1,6 +1,6 @@
 
 [vectorizer]
-vectorizer = knext.common.vectorizer.OpenAIVectorizer
+vectorizer = kag.common.vectorizer.OpenAIVectorizer
 model = bge-m3
 api_key = EMPTY
 base_url = http://127.0.0.1:11434/v1
